+++ ProcessMaker/Http/Controllers/AboutController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProcessMaker\Http\Controllers\Admin;
+namespace ProcessMaker\Http\Controllers;
 
 use Illuminate\Http\Request;
 use ProcessMaker\Http\Controllers\Controller;
@@ -16,6 +16,6 @@ class AboutController extends Controller
      */
     public function index()
     {
-        return view('admin.about.index');
+        return view('about.index');
     }
 }
