+++ ProcessMaker/Http/Middleware/GenerateMenus.php
@@ -55,13 +55,6 @@ class GenerateMenus
                 'icon' => 'fa-globe',
                 'id' => 'homeid'
           ]);
-
-          $submenu = $menu->add(__('menus.sidebar_admin.system_information'));
-          $submenu->add(__('menus.sidebar_admin.app_version_details'), [
-                'route' => 'about.index',
-                'icon' => 'fa-desktop',
-                'id' => 'homeid'
-          ]);
           $submenu->add(__('menus.sidebar_admin.queue_management'), [
                 'route' => 'horizon.index',
                 'icon' => 'fa-infinity',
@@ -147,6 +140,14 @@ class GenerateMenus
             'header' => false,
             'route' => 'home',
             'icon' => 'fa-info',
+            'img' => '',
+            'id' => 'dropdownItem'
+          ],
+          [
+            'label' => __('About'),
+            'header' => false,
+            'route' => 'about.index',
+            'icon' => 'fa-question',
             'img' => '',
             'id' => 'dropdownItem'
           ],
