+++ resources/views/about/index.blade.php
@@ -10,17 +10,17 @@
 
 @section('content')
  <div class="container">
-    <h1>About ProcessMaker</h1>
+    <h1>{{__('About ProcessMaker 4')}}</h1>
     <div class="row">
       <div class="col-8">
         <div class="card card-body">
         <img class="about-logo" src="/img/md-blk-logo.png">
         <hr>
-        <div>ProcessMaker v4.0</div>
+        <div>{{__('ProcessMaker v4.0')}}</div>
         <hr>
-        <a href="https://github.com/ProcessMaker/bpm/issues">Report an issue  <i class="fas fa-caret-right fa-lg float-right mr-1"></i></a>
+        <a href="https://github.com/ProcessMaker/bpm/issues">{{__('Report an issue')}}  <i class="fas fa-caret-right fa-lg float-right mr-1"></i></a>
         <hr>
-        <a href="https://github.com/ProcessMaker/bpm">Get Help  <i class="fas fa-caret-right fa-lg float-right mr-1"></i></a>
+        <a href="https://github.com/ProcessMaker/bpm">{{__('Get Help')}}  <i class="fas fa-caret-right fa-lg float-right mr-1"></i></a>
         <hr>
       </div>
     </div>
